Moved the htmlchars and htmlchars_decode functions out of clean_request.php and into func.php, where I also added them to the $func array as well.
Changes to database.php were made so if database connection failed, displaying the error was called in the connect method, not by database.php

SQLite class added, quite a bit done, no result class yet, though, and also the insert method is not done either.

A bit of stuff added to the upload class.