The API class has been completely recoded. The methods run_hook and apply_filter has been renamed to run_hooks and apply_filters, respectively.
Action, subaction and request parameter support have been completely removed, in favor of events, which allows you to specify all the aforementioned in one. However, wildcards are not supported, though in the next update I hope to add it!

As for the mass edits, this has been done because all API calls have been changed, of course!