Moved the Control Panel text under the sites name in the Admin CP.
Added an array_insert function to compat.php, to of course, insert an item at the specified position in an array.

Added some accessors to the Form class, along with the ability to specify a position of where the field is added.

$settings wasn't global'd in the init_session function in session.php.

Some more work done on the Table class.