The Table class was redone in part to allow a function to be set which will return a set of data, rather than requiring the data to come from SQL only. Not completely finished, but for the most part, it is ;)
Some bugs in the API class fixed, had to do with events. More than likely it will have to be redone as it sees the id=* in action=admin&sa=members_managed&id=* as the same of id under another sa (sub action) than members_manage, when it shouldn't!

timeformat function now works.