Fragmented the control panel into more files, doubt there will be any performance increase, but hey, less code to search through ;-) (Fixes issue #56).
PHP 5.3.2 removed the automatic unlocking of a file's resource handling, so I went through and added flock(resource, LOCK_UN); where needed.

Moved the recursive_unlink function from admin_update.php to compat.php.

The RSS class has been completed (in regards to issue #55) and tested. Next: Atom class.

A new function, verify_request, has been added, which checks the specified method (get, post or request) to see if sid (members session id) is supplied and valid. Useful for stopping CSRF, you know, making an administrator or someone do something you didn't want in cases that the Tokens class is a little overkill.

The Tar class will now do some basic testing to see if the supplied file is actually a tarball (just checks to see if the last 512 bytes are NUL-bytes, as that is *about* the only way to do it).

Added accessors to the Theme class, but also removed the json_encode call in the add_js_var method, as when the js variable would be accessed via return_js_var could return unexpected results. The call to json_encode has been moved to the Implemented_Theme class.

Added a new class, XML, which reads and parses XML data, tested and what not. It uses XML Parser, and it changes the format of the parsed data a bit, but not much.

Added a theme.xml file, as I plan to soon remove the ini files for information about themes/plugins and replace it with XML files (which look nicer, usually).