Converted the rest of the main component of the system over to the new theme system.
Along with that, I made a few changes to the Theme class in order to allow the current theme to be changed on the fly with a new attribute called themedir which simply contains the current directory of where the header.template.php and footer.template.php files are located.

I deleted the Admin_Theme class, as I have converted the admin theme over to the new setup.

I did not convert over much of the control panel component of the system, just the authentication page and the admin home page for right now.