Converted over more pages to the new theme system.
I have converted over the following components of the control panel:
- System Settings
- Themes
- Update 
- About
- Errors

I have yet to convert any of the member management or plugin management components yet.

Just as a note, the templates for the control panel components are kind of ugly, and contain a lot of PHP code and other operations. I am not sure if I should move these out, or leave it the way it is, as it might be expected to be somewhat ugly...