I can't believe I am saying this, but plugin management (adding, activating, deactivating, deleting and updating) is complete. Yay!
I made a couple pretty big changes, as I have created the Component class which handles the installing and updating of components (plugins and themes). I did this for maximum code reuse :-) The Component class has taken place of a couple of functions, such as checking the status of a component and interpreting the retrieved status.

- admin_plugins_get_message and plugin_check_status have been removed and replaced with new methods in the Component class.
! The tasks ran by users browsing the website are now not started until the page has been loaded.